--- a/event_test.go
+++ b/event_test.go
@@ -11,8 +11,8 @@ func TestParseEventJoin(t *testing.T) {
 	// TODO: Use new log to fix
 	a := assert.New(t)
 	s := &Segment{
-		Line: "[UnityCrossThreadLogger]1/8/2019 2:11:59 PM",
-		Text: `
+		Line: []byte(`[UnityCrossThreadLogger]1/8/2019 2:11:59 PM`),
+		Text: []byte(`
 <== Event.Join(57)
 {
   "Id": "6c236ae7-81ff-4024-a836-cc055194fafe",
@@ -22,7 +22,7 @@ func TestParseEventJoin(t *testing.T) {
   "CurrentModule": "PayEntry",
   "CardPool": null,
   "CourseDeck": null
-}`,
+}`),
 	}
 	joined, err := s.ParseEventJoin()
 	a.Nil(err)
@@ -35,8 +35,8 @@ func TestParsePayEntry(t *testing.T) {
 	t.Skip()
 	a := assert.New(t)
 	s := &Segment{
-		Line: "[UnityCrossThreadLogger]1/8/2019 2:11:59 PM",
-		Text: `
+		Line: []byte(`[UnityCrossThreadLogger]1/8/2019 2:11:59 PM`),
+		Text: []byte(`
 <== Event.PayEntry(58)
 {
   "Id": "6c236ae7-81ff-4024-a836-cc055194fafe",
@@ -82,7 +82,7 @@ func TestParsePayEntry(t *testing.T) {
     "lockedForEdit": false,
     "isValid": false
   }
-}`,
+}`),
 	}
 	payed, err := s.ParseEventPayEntry()
 	a.Nil(err)
@@ -94,8 +94,8 @@ func TestParsePayEntry(t *testing.T) {
 func TestParseEventGetPlayerCourse(t *testing.T) {
 	a := assert.New(t)
 	s := &Segment{
-		Line: "[UnityCrossThreadLogger]1/8/2019 2:12:00 PM",
-		Text: `
+		Line: []byte(`[UnityCrossThreadLogger]1/8/2019 2:12:00 PM`),
+		Text: []byte(`
 <== Event.GetPlayerCourse(63)
 {
   "Id": "6c236ae7-81ff-4024-a836-cc055194fafe",
@@ -143,7 +143,7 @@ func TestParseEventGetPlayerCourse(t *testing.T) {
   }
 }
 new prize bar state is: PrizeDisplay
-`,
+`),
 	}
 	pc, err := s.ParseJoinedEvent()
 	a.Nil(err)