--- a/segment.go
+++ b/segment.go
@@ -1,6 +1,7 @@
 package gathering
 
 import (
+	"bytes"
 	"encoding/json"
 	"regexp"
 	"time"
@@ -49,7 +50,7 @@ var segmentTypeChecks = map[SegmentType]*regexp.Regexp{
 	EventGetPlayerCourse:              regexp.MustCompile(`<==\sEvent\.GetPlayerCourse\(\d+\)`),
 	MatchStart:                        regexp.MustCompile(`Incoming\sEvent\.MatchCreated`),
 	MatchEnd:                          regexp.MustCompile(`DuelScene\.GameStop`),
-	MatchEvent:                        regexp.MustCompile(`GREMessageType_GameStateMessage|GameStateType_Diff`),
+	MatchEvent:                        regexp.MustCompile(`"GREMessageType_GameStateMessage"|GameStateType_Diff`),
 	EventDeckSubmit:                   regexp.MustCompile(`<==\sEvent\.DeckSubmit\(\d+\)`),
 	CrackBooster:                      regexp.MustCompile(`<==\sPlayerInventory\.CrackBoostersV3\(\d+\)`),
 	InventoryRankUpdated:              regexp.MustCompile(`Incoming\sRank\.Updated`),
@@ -57,17 +58,42 @@ var segmentTypeChecks = map[SegmentType]*regexp.Regexp{
 	IncomingInventoryUpdate:           regexp.MustCompile(`Incoming\sInventory\.Updated`),
 }
 
+var cleaners = []*regexp.Regexp{
+	regexp.MustCompile(`<<<<<<<<<<.*`),
+	regexp.MustCompile(`\[\w.*`),
+	regexp.MustCompile(`\dx[\d\w]+.*`),
+	regexp.MustCompile(`(?m)ZoneTransferUXEvent.*`),
+	regexp.MustCompile(`(?sm)BIError - GRE.Notification:.*`),
+}
+var clean = []byte(`$1.$2`)
+
+var trimLeft = func(r rune) bool {
+	return r != '{' && r != '['
+}
+var trimRight = func(r rune) bool {
+	return r != '}' && r != ']'
+}
+
 // Segment is a piece of the log
 type Segment struct {
 	LoggerType  LoggerType
 	Time        *time.Time
 	SegmentType SegmentType
-	Text        string
+	Text        []byte
 	Range       []int
-	Line        string
+	Line        []byte
 }
 
 // JSON parses the text as JSON
-func (s *Segment) JSON(i interface{}) error {
-	return json.Unmarshal([]byte(s.Text), i)
+func (s *Segment) JSON(v interface{}) error {
+	return json.Unmarshal(s.Text, v)
+}
+
+func stripNonJSON(b []byte) []byte {
+	for _, c := range cleaners {
+		b = c.ReplaceAll(b, clean)
+	}
+	b = bytes.TrimLeftFunc(b, trimLeft)
+	b = bytes.TrimRightFunc(b, trimRight)
+	return b
 }