Performance has always been pretty abysmal with the parser, but there were more pressing issues to work on. However, after playing a bunch of games and the parser eating too much memory and CPU, it was time to address it.

Working from [here](https://blog.golang.org/profiling-go-programs) and parsing a `133mb` log file 5 times.

How much?

```
379.51u 6.16s 220.90r 2091420kB ./gathering -file=big-profile-log.txt -token=token
```
It took 220.90 seconds to run and used a massive `2091.42mb` of RAM.

Okay. That's a problem.

Ran the Go profiling tools and discovered that the source of memory issues was actually here:
https://github.com/gathering-gg/parser/compare/profiled?expand=1#diff-308aaa719425f4333033c97cfb3b9868L46

Each concatenation was causing a lot of memory allocation and re-allocation. Go has trouble with string concatenation. http://herman.asia/efficient-string-concatenation-in-go

That's okay. As seen in the PR, we switched over to `bytes.Buffer` method, which reuses the buffer. We also changed all methods to use `[]byte` methods and removed as much `string` usage as possible.

How do things look now?

```
37.00u 0.88s 37.54r 351044kB ./gathering -file=profile-log.txt -token=token
```
Parsing took `37.54s` and only used `351.044mb` of memory! That's an average of 7.5s to parse the log file, a performance increase of 600%. Pretty good!